modify wastebasket comment feature

+++ src/main/java/org/themselves/alber/service/WastebasketService.java
@@ -37,7 +37,7 @@ public class WastebasketService {
     private final S3Uploader s3Uploader;
 
     @Transactional
-    public void addWastebasket(Wastebasket wastebasket, User user, List<Long> imageList) {
+    public void addWastebasket(Wastebasket wastebasket, User user, List<String> imageList) {
 
         //이미지가 3개 이상이면??
         if(imageList.size() > 3) {
@@ -57,9 +57,9 @@ public class WastebasketService {
         pin.setWastebasket(newWastebaseket);
         pinRepository.save(pin);
 
-        for(Long imageId : imageList) {
+        for(String imageId : imageList) {
 
-            Optional<Image> op_image = imageRepository.findById(imageId);
+            Optional<Image> op_image = imageRepository.findById(Long.parseLong(imageId));
             if(!op_image.isPresent())
                 throw new CustomException(StatusCode.IMAGE_NOT_FOUND);
             //쓰레기통 이미지 저장
@@ -87,11 +87,12 @@ public class WastebasketService {
         return wastebasketRepository.findAll(pageable);
     }
 
-    @Transactional
-    public void setWastebasketOne(Wastebasket wastebasket, MultipartFile[] files) {
 
-        //파일이 3개 이상이면??
-        if (files.length > 3) {
+    @Transactional
+    public void setWastebasketOne(Wastebasket wastebasket, User user, List<String> imageList) {
+
+        //이미지가 3개 이상이면??
+        if(imageList.size() > 3) {
             throw new CustomException(StatusCode.FILE_TO_MUCH_ERROR);
         }
 
@@ -102,53 +103,33 @@ public class WastebasketService {
 
         updateWastebasket.get().setBoxName(wastebasket.getBoxName());
         updateWastebasket.get().setAreaDesc(wastebasket.getAreaDesc());
-        updateWastebasket.get().setAreaSi(wastebasket.getAreaSi());
-        updateWastebasket.get().setAreaGu(wastebasket.getAreaGu());
-        updateWastebasket.get().setAreaDong(wastebasket.getAreaDong());
-        updateWastebasket.get().setLatitude(wastebasket.getLatitude());
         updateWastebasket.get().setLongitude(wastebasket.getLongitude());
+        updateWastebasket.get().setLatitude(wastebasket.getLatitude());
+        wastebasketRepository.save(updateWastebasket.get());
 
-        //이전 이미지 데이터 삭제
-        for (WastebasketImage wastebasketImage : updateWastebasket.get().getImageList()){
-            new File(wastebasketImage.getImage().getUrl()).delete();
-            wastebasketImageRepository.delete(wastebasketImage);
-            imageRepository.delete(wastebasketImage.getImage());
+        //쓰레기통 이미지 저장
+        for(String imageId : imageList) {
 
-        }
-//        updateWastebasket.get().setImageList(new ArrayList<>());
+            Optional<Image> op_image = imageRepository.findById(Long.parseLong(imageId));
+            if(!op_image.isPresent())
+                throw new CustomException(StatusCode.IMAGE_NOT_FOUND);
 
-        Path rootPath = FileUtil.directoryExistAndCreate("image");
-        Path childPath = FileUtil.directoryExistAndCreate(rootPath.toString()
-                + File.separator + "wastebasket_" + updateWastebasket.get().getId());
-
-
-        for (MultipartFile file : files) {
-
-            //파일 형식이 이미지가 아니면?
-            if (!FileUtil.fileEqImage(file))
-                throw new CustomException(StatusCode.FILE_NOT_IMAGE_ERROR);
-
-            String url;
-            try {
-                url = s3Uploader.upload(file);
-            } catch (IOException e) {
-                throw new CustomException(StatusCode.FILE_CREATE_ERROR);
-            }
-
-            //이미지저장
-            Image image = new Image();
-
-            image.setUrl(url);
-            imageRepository.save(image);
-
-            //쓰레기통 이미지 저장
             WastebasketImage wi = new WastebasketImage();
-            wi.setImage(image);
+            wi.setImage(op_image.get());
             wi.setWastebasket(updateWastebasket.get());
             wastebasketImageRepository.save(wi);
-
         }
+
+
     }
+
+
+
+
+
+
+
+
 
     @Transactional
     public void delWastebasket(Long id, User user) {
@@ -160,13 +141,15 @@ public class WastebasketService {
         Pin pin = wastebasket.get().getPinList().get(0);
 
         //댓글이 있다면
+        if(wastebasket.get().getWastebasketCommentList().size() != 0 )
+            throw new CustomException(StatusCode.WASTEBASKET_NOT_DELETE_EXIST_COMMENTS);
 
         //본인이 올린 쓰레기통이 아니라면
         if(!user.getId().equals(pin.getUser().getId()))
             throw new CustomException(StatusCode.WASTEBASKET_NOT_SAME_USER);
 
         //쓰레기통이미지/이미지삭제
-        for (WastebasketImage wastebasketImage : wastebasket.get().getImageList()){
+        for (WastebasketImage wastebasketImage : wastebasket.get().getWastebasketImageList()){
             new File(wastebasketImage.getImage().getUrl()).delete();
             wastebasketImageRepository.delete(wastebasketImage);
             imageRepository.delete(wastebasketImage.getImage());
